requirements.txt
py-cord
httpx
python-dotenv
pytesseract
Pillow